Reverse the update logic
Before:

A table mapped from minutes to IDs of elements to highlight,
and the highlight class was added to each of these elements.

Now:

Each element knows for each interval whether it should be
highlighted or not by giving it a list of CSS class names.
The update itself is then done by dynamically changing the name
of the highlighted class in the style sheet.